Include Deploy on Platform.sh button
Since the README already includes instructions to deploy this project on Platform.sh, it makes sense to include the shorter way.  I put the button where it seemed to make sense logically but if you think it works better elsewhere feel free to move it.  (The UTM tags are just so we know where the link is coming from and who to give credit to.)